backend/requirements.txt
asarPy==1.0.1
asgiref==3.8.1
Django==5.0.6
django-cors-headers==4.4.0
django-guest-user==0.5.5
djangorestframework==3.15.2
sqlparse==0.5.0
